added radio input generator
