+++ RazorForms/Generators/Inputs/ICheckInputGenerator.cs
@@ -1,14 +1,5 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Razor.TagHelpers;
-using RazorForms.Options;
-using RazorForms.TagHelpers;
+﻿namespace RazorForms.Generators.Inputs;
 
-namespace RazorForms.Generators.Inputs;
-
-public interface ICheckInputGenerator : IOutputGenerator<IFormComponentOptions>
+public interface ICheckInputGenerator : ICheckRadioInputGenerator
 {
-	Task<TagHelperOutput> GenerateOutput(TagHelperContext context,
-	                                     RazorFormsTagHelperBase helper,
-	                                     TagHelperAttributeList? attributes = null,
-	                                     TagHelperContent? childContent = null);
 }
