+++ RazorForms/Generators/Inputs/IRadioInputGenerator.cs
@@ -1,0 +1,5 @@
+﻿namespace RazorForms.Generators.Inputs;
+
+public interface IRadioInputGenerator : ICheckRadioInputGenerator
+{
+}
