+++ RazorForms/Generators/Inputs/RadioInputGenerator.cs
@@ -1,0 +1,10 @@
+﻿namespace RazorForms.Generators.Inputs;
+
+public class RadioInputGenerator : CheckRadioInputGeneratorBase, IRadioInputGenerator
+{
+	/// <inheritdoc />
+	public RadioInputGenerator()
+	{
+		TypeAttributeValue = "radio";
+	}
+}
